Add an ability to set the cookie to insecure mode also in production.
Conduit Suite is using production Lando UI by default. Suite is running on
http and cookies needs to stay insecure.

If `SESSION_COOKIE_SECURE` environment variable was set to 0 uwsgi was
reading it as a string - `"0"`. This is now changed to int.

Talisman is taking control od the `SESSION_COOKIE_SECURE` config and
sets it to `True` if app is not in debug mode. This patch is making the
uwsgi to read the `DEBUG` environment variable and passes it to the app.

Also we're now logging most of the session configs.