WIP - landing: Add an ability to see and land secure revisions.
Summary:
The user is providing the Phabricator API Token manually in the Navbar modal.
The token is posted to the `/settings` endpoint.
The browser is storing its value in the httponly cookie.
Lando UI backend is passing the token to the Lando API in request's header.

Test Plan: Attempt to land a secure revision.

Reviewers: imadueme smacleod!

Tags: #secure-revision

Bug #: 1443704

Differential Revision: https://phabricator.services.mozilla.com/D2561